Fix dynamic bridge start-up delay
Signed-off-by: Chris Elston <chris.elston@sancloud.com>

The first packet sent by a dynamic bridge was failing because it was
sent before the socket was fully established and the OS returned
-EAGAIN. We had to wait for the next PINGREQ to cause the initial packet
to be sent, and therefore bridge startup was delayed by the
configured keepalive timeout (default 60 seconds).

This change adds the new bridge's output socket to the list managed by
the mux, and we now send the initial CONNECT as soon as the output
socket becomes available.